MAJOR UPGRADE VERSION FOR MULTI-(EQUIVALENT)SITE CALCULATIONS.
From this commit on also the interaction tensor will be rotated when the impurity is solved in a different basis with respect to the lattice.
It has been thoroughly tested that the rotation routine leaves unchanged the interaction if the latter has the Kanamori symmetry from the beginning.

Results on LiTi2O4 shows how this improves greatly the convergence accuracy on the interaction.

The Fermionic part remains unchanged with respect to the previous versions.